// src/app/core/models/onboarding/country-response.ts
export interface CountryResponse {

  id: number;

  uuid: string;

  libelle: string;

  code: string;

  iso2: string;

  iso3: string;

  phoneCode: string;

  currencyId: number;

  currencyCode: string;

  currencyLibelle: string;

  languageId: number;

  languageLocale: string;

  timeZoneId: number;

  timeZoneZoneId: string;
}