// src/app/core/models/onboarding/onboarding-organization-request.ts
export interface OnboardingOrganizationRequest {

  organizationTypeUuid: string;

  libelle: string;

  code: string;

  countryUuid: string;

  regionUuid: string;

  departmentUuid: string;

  sigle: string;

  adresse: string;

  boitePostale: string;

  telephone: string;

  mobile: string;

  email: string;

  siteWeb: string;

  logoFileUuid: string;

  description: string;
}